Fix html-editor not updating to reflect content after cell shifted
Pass a key that is unique for each cell. It is based off the entity self href or level + self href (in the case of criterionCell as they are always 0, 1, 2...)
When the cell is shifted and the html-editor is re-used, the shifted cell will have a new key value. This will notify the html-editor that it needs to re-set content to reflect its new value.